Update index.html
Edited lead text a bit - from 'This is a page dedicated to helping writers by giving tips in regards to how to come up with an idea to a literary text, how to develop the idea and increase the quality of the text and then end up with a finished product. Here you can find collected information from other sources, quotations, information based on experience etc.' to 'This is a page dedicated to helping writers. Here you can find tips on how to come up with an idea, how to develop it and then eventually end up with a finished product. We aim to provide you with versatile information based on qualitative sources as well as personal experience.'